Add support for sub-second intervals
Interval is no more limited to "seconds":

```shell
./dnsping.py -i 0.2 -c 50 --flags  -t AAAA -s 9.9.9.9 ripe.net
```

Fixes #89

+++ dnsping.py
@@ -158,7 +158,7 @@ def main():
         elif o in ("-w", "--wait"):
             timeout = int(a)
         elif o in ("-i", "--interval"):
-            interval = int(a)
+            interval = float(a)
         elif o in ("-t", "--type"):
             rdatatype = a
         elif o in ("-T", "--tcp"):
